BUGFIX: Support multiple variable formats
Currently the only format the HTTP endpoint support for
the variables is a JSON-serialized string:

```
{
	"query": "<some-query>",
	"variables": "{\"id\": \"123\"}"
    // ...
}
```

With this change, unserialized JSON objects are supported, too:

```
```
{
	"query": "<some-query>",
	"variables": {
       "id": "123"
    }
    // ...
}
```

Note: I couldn't find anything about this in the GraphQL spec.
The GraphiQL IDE sends the variables serialized, but other endpoints
seem to support both formats.