BUGFIX: Describe and catch circular dependencies
The latest version of the `webonyx/graphql-php` package
does not allow GraphQL types to be closures.
Instead the "fields" configuration has to be a closure
in order to realize types with circular dependencies.

See https://webonyx.github.io/graphql-php/type-system/object-types/#recurring-and-circular-types

This fix adjusts the `TypeResolver` accordingly so that
it throws an exception with further details if a circular
dependency was configured without closure.